add last changes with fast commit

+++ src/database/migrations/create_role_users_table.php
@@ -13,7 +13,7 @@ class CreateRoleUsersTable extends BaseMigration
         $table->engine = 'InnoDB';
 
         $table->bigIncrements('id');
-        $table->bigInteger('role_id')->unsigned();
+        $table->integer('role_id')->unsigned();
         $table->bigInteger('user_id')->unsigned();
         $table->timestamps();
         $table->foreign('user_id')
